requirements.txt
nltk
numpy
emoji
contractions
pyspellchecker